Add coroutine dependency, move interfaces to separated folder
