+++ app/src/main/java/com/app/music_app/note_player/instruments/VirtualPiano.kt
@@ -1,5 +1,6 @@
 package com.app.music_app.note_player.instruments
 
+import com.app.music_app.note_player.interfaces.AbstractInstrument
 import com.musiclib.Alteration
 import com.musiclib.notes.Note
 import com.musiclib.notes.NoteName
